Optimize ControlPad::read, simplify tests, and remove benchmark
